v0.6稳定版 4_25
v0.6稳定版

+++ Core/User/Inc/fes.h
@@ -34,6 +34,6 @@ void merge_stimulate_parameter(UART_HandleTypeDef *huart,int amplitude,int chann
 void stim_search(UART_HandleTypeDef *huart);
 void stim_start(UART_HandleTypeDef *huart);
 void stim_stop(UART_HandleTypeDef *huart);
-void stimulate(UART_HandleTypeDef *huart,float amp,int channel);
+void stimulate(UART_HandleTypeDef *huart,double amp,int channel);
 
 #endif
